fishr: Access the Fish2 FlyEM Dataset

A thin wrapper around the malevnc package providing access to the fish2 larval zebrafish EM dataset being proofread at Janelia Research Campus as part of a collaboration with the Engert / Lichtman labs at Harvard and Google Research. Currently provides convenience functions for reading meshes and setting dataset options.

Package Options

fishr.dataset The name of the active fish2 dataset (default "fish2"). Set by choose_fish_dataset.

The following malevnc.* options are set by choose_fish or with_fish to point the malevnc package at the fish2 servers:

malevnc.server DVID server URL (set via Clio lookup).

malevnc.rootnode DVID root node UUID (set via Clio lookup).

malevnc.neuprint neuprint server URL (fallback: https://neuprint-fish2.janelia.org).

malevnc.neuprint_dataset neuprint dataset name (fallback: "fish2").

See also

Author

Maintainer: Gregory Jefferis jefferis@gmail.com (ORCID)

Examples

# \donttest{
options()[grepl("^fishr", names(options()))]
#> $fishr.dataset
#> [1] "fish2"
#> 
# }
if (FALSE) { # \dontrun{
with_fish(read_fish_meshes(12345))
} # }