Read meshes for fish2 body ids

Fetches pre-computed neuroglancer meshes from the fish2 DVID server.

Usage

read_fish_meshes(ids, node = "neutu", units = c("nm", "raw", "microns"), ...)

Arguments

ids

One or more body ids (numeric or character), or a query string accepted by fish_ids.

node

The DVID node (UUID) to query. The default value of 'neutu' uses the active neutu node i.e. normally the most up to date.

units

One of "nm" (default), "raw" (nominal voxel size), or "microns".

...

Additional arguments passed to httr::GET.

Value

A neuronlist containing one or more mesh3d objects.

Details

It seems that the nominal voxel dimensions of the fish2 dataset are 16 x 16 x 15 nm. Neuron meshes are already returned from DVID in physical nm scale but this voxel dimension is required to turn raw neuroglancer coordinates into nm. This means that if you click on a neuron in neuroglancer, the recorded location will match if you have chosen `units="nm"`. Alternatively you must multiply point locations from neuroglancer by c(16,16,15) to obtain points in nm space.

See also

Examples

if (FALSE) { # \dontrun{
# pick 10 RGC meshes to read in
rgcs <- read_fish_meshes(fish_ids('RGC')[1:10])
plot3d(rgcs)
} # }